// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
// ...
export interface LicensingOption {
    id: string;
    terms: string;
    name: string;
    description: string;
    price: bigint;
}
// ...
export enum MediaType {
    music = "music",
    video = "video",
    text = "text",
    image = "image"
}